pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "medical-audit-v2"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.3",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "cryptography>=42",
    "structlog>=24",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "pdfplumber>=0.11",
    "ocrmypdf>=16",
    "playwright>=1.44",
    "google-api-python-client>=2.130",
    "google-auth>=2.30",
    "aiofiles>=23",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["ruff", "mypy", "pytest", "pytest-asyncio", "httpx"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "core*", "seeds*"]

[tool.ruff]
line-length = 120
target-version = "py311"
